+++ HISTORY.rst
@@ -3,18 +3,19 @@
 History
 -------
 
-0.1.0 (2014-01-11)
+0.1.0 (01 Nov 2014)
 ---------------------
 
 * First release of APSG
 
-0.2.0 (near future)
+0.2.0 (09 Dec 2014)
 ---------------------
 
 * new StereoNet class for Schmidt projection
+* Quick plot when data are passed to StereoNet class instantiation
 * mplstereonet dependency depreceated
 
-* new Pair and Fault classes to manipulate attached data
+* new Pair and Fault classes to manipulate paired data (full support in future)
 * new uniform_lin and uniform_fol Group methods
 * abs for Group implemented to calculate euclidean norms
 * new Group method normalized